+++ tests/corelib/range_error_test.dart
@@ -48,7 +48,8 @@ void testToString() {
       for (var re in [
         new ArgumentError.value(value, name, message),
         new RangeError.value(value, name, message),
-        new RangeError.index(value, [], name, message),
+        new IndexError.withLength(value, 0,
+            indexable: [], name: name, message: message),
         new RangeError.range(value, 0, 24, name, message)
       ]) {
         var str = re.toString();